refactor : Empathy 기능 좋아요 카운트 추가

+++ src/main/java/com/hamtaro/sunflowerplate/repository/EmpathyRepository.java
@@ -13,6 +13,6 @@ public interface EmpathyRepository extends JpaRepository<EmpathyEntity , Long> {
 
        void deleteByMemberEntityAndReviewEntity(MemberEntity memberEntity , ReviewEntity reviewEntity);
 
-       long countByReviewEntity(ReviewEntity reviewEntity);
+       int countByReviewEntity(ReviewEntity reviewEntity);
 
 }
